@@ -14,12 +14,8 @@ The MSBuild schemas were also considered.  They are typically found in the
 MSBuild install directory, e.g. c:\Program Files (x86)\MSBuild
 """
 
-from __future__ import print_function
-
-from gyp import string_types
-
-import sys
 import re
+import sys
 
 # Dictionaries of settings validators. The key is the tool name, the value is
 # a dictionary mapping setting names to validation functions.
@@ -36,7 +32,7 @@ _msvs_to_msbuild_converters = {}
 _msbuild_name_of_tool = {}
 
 
-class _Tool(object):
+class _Tool:
     """Represents a tool used by MSVS or MSBuild.
 
   Attributes:
@@ -68,7 +64,7 @@ def _GetMSBuildToolSettings(msbuild_settings, tool):
     return msbuild_settings.setdefault(tool.msbuild_name, {})
 
 
-class _Type(object):
+class _Type:
     """Type of settings (Base class)."""
 
     def ValidateMSVS(self, value):
@@ -110,11 +106,11 @@ class _String(_Type):
     """A setting that's just a string."""
 
     def ValidateMSVS(self, value):
-        if not isinstance(value, string_types):
+        if not isinstance(value, str):
             raise ValueError("expected string; got %r" % value)
 
     def ValidateMSBuild(self, value):
-        if not isinstance(value, string_types):
+        if not isinstance(value, str):
             raise ValueError("expected string; got %r" % value)
 
     def ConvertToMSBuild(self, value):
@@ -126,11 +122,11 @@ class _StringList(_Type):
     """A settings that's a list of strings."""
 
     def ValidateMSVS(self, value):
-        if not isinstance(value, string_types) and not isinstance(value, list):
+        if not isinstance(value, (list, str)):
             raise ValueError("expected string list; got %r" % value)
 
     def ValidateMSBuild(self, value):
-        if not isinstance(value, string_types) and not isinstance(value, list):
+        if not isinstance(value, (list, str)):
             raise ValueError("expected string list; got %r" % value)
 
     def ConvertToMSBuild(self, value):
@@ -195,7 +191,7 @@ class _Enumeration(_Type):
     def __init__(self, label_list, new=None):
         _Type.__init__(self)
         self._label_list = label_list
-        self._msbuild_values = set(value for value in label_list if value is not None)
+        self._msbuild_values = {value for value in label_list if value is not None}
         if new is not None:
             self._msbuild_values.update(new)
 
@@ -342,7 +338,7 @@ def _ConvertedToAdditionalOption(tool, msvs_name, flag):
         if value == "true":
             tool_settings = _GetMSBuildToolSettings(msbuild_settings, tool)
             if "AdditionalOptions" in tool_settings:
-                new_flags = "%s %s" % (tool_settings["AdditionalOptions"], flag)
+                new_flags = "{} {}".format(tool_settings["AdditionalOptions"], flag)
             else:
                 new_flags = flag
             tool_settings["AdditionalOptions"] = new_flags
@@ -536,14 +532,14 @@ def _ValidateSettings(validators, settings, stderr):
                         tool_validators[setting](value)
                     except ValueError as e:
                         print(
-                            "Warning: for %s/%s, %s" % (tool_name, setting, e),
+                            f"Warning: for {tool_name}/{setting}, {e}",
                             file=stderr,
                         )
                 else:
                     _ValidateExclusionSetting(
                         setting,
                         tool_validators,
-                        ("Warning: unrecognized setting %s/%s" % (tool_name, setting)),
+                        (f"Warning: unrecognized setting {tool_name}/{setting}"),
                         stderr,
                     )
 
