@@ -2,7 +2,6 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-from __future__ import print_function
 
 import collections
 import copy
@@ -20,10 +19,7 @@ import gyp.msvs_emulation
 import gyp.MSVSUtil as MSVSUtil
 import gyp.xcode_emulation
 
-try:
-    from cStringIO import StringIO
-except ImportError:
-    from io import StringIO
+from io import StringIO
 
 from gyp.common import GetEnvironFallback
 import gyp.ninja_syntax as ninja_syntax
@@ -76,7 +72,7 @@ def StripPrefix(arg, prefix):
 
 def QuoteShellArgument(arg, flavor):
     """Quote a string such that it will be interpreted as a single argument
-  by the shell."""
+    by the shell."""
     # Rather than attempting to enumerate the bad shell characters, just
     # allow common OK ones and quote anything else.
     if re.match(r"^[a-zA-Z0-9_=.\\/-]+$", arg):
@@ -88,7 +84,7 @@ def QuoteShellArgument(arg, flavor):
 
 def Define(d, flavor):
     """Takes a preprocessor define and returns a -D parameter that's ninja- and
-  shell-escaped."""
+    shell-escaped."""
     if flavor == "win":
         # cl.exe replaces literal # characters with = in preprocessor definitions for
         # some reason. Octal-encode to work around that.
@@ -99,32 +95,32 @@ def Define(d, flavor):
 def AddArch(output, arch):
     """Adds an arch string to an output path."""
     output, extension = os.path.splitext(output)
-    return "%s.%s%s" % (output, arch, extension)
+    return f"{output}.{arch}{extension}"
 
 
-class Target(object):
+class Target:
     """Target represents the paths used within a single gyp target.
 
-  Conceptually, building a single target A is a series of steps:
+    Conceptually, building a single target A is a series of steps:
 
-  1) actions/rules/copies  generates source/resources/etc.
-  2) compiles              generates .o files
-  3) link                  generates a binary (library/executable)
-  4) bundle                merges the above in a mac bundle
+    1) actions/rules/copies  generates source/resources/etc.
+    2) compiles              generates .o files
+    3) link                  generates a binary (library/executable)
+    4) bundle                merges the above in a mac bundle
 
-  (Any of these steps can be optional.)
+    (Any of these steps can be optional.)
 
-  From a build ordering perspective, a dependent target B could just
-  depend on the last output of this series of steps.
+    From a build ordering perspective, a dependent target B could just
+    depend on the last output of this series of steps.
 
-  But some dependent commands sometimes need to reach inside the box.
-  For example, when linking B it needs to get the path to the static
-  library generated by A.
+    But some dependent commands sometimes need to reach inside the box.
+    For example, when linking B it needs to get the path to the static
+    library generated by A.
 
-  This object stores those paths.  To keep things simple, member
-  variables only store concrete paths to single files, while methods
-  compute derived values like "the last output of the target".
-  """
+    This object stores those paths.  To keep things simple, member
+    variables only store concrete paths to single files, while methods
+    compute derived values like "the last output of the target".
+    """
 
     def __init__(self, type):
         # Gyp type ("static_library", etc.) of this target.
@@ -163,7 +159,7 @@ class Target(object):
 
     def UsesToc(self, flavor):
         """Return true if the target should produce a restat rule based on a TOC
-    file."""
+        file."""
         # For bundles, the .TOC should be produced for the binary, not for
         # FinalOutput(). But the naive approach would put the TOC file into the
         # bundle, so don't do this for bundles for now.
@@ -173,19 +169,19 @@ class Target(object):
 
     def PreActionInput(self, flavor):
         """Return the path, if any, that should be used as a dependency of
-    any dependent action step."""
+        any dependent action step."""
         if self.UsesToc(flavor):
             return self.FinalOutput() + ".TOC"
         return self.FinalOutput() or self.preaction_stamp
 
     def PreCompileInput(self):
         """Return the path, if any, that should be used as a dependency of
-    any dependent compile step."""
+        any dependent compile step."""
         return self.actions_stamp or self.precompile_stamp
 
     def FinalOutput(self):
         """Return the last output of the target, which depends on all prior
-    steps."""
+        steps."""
         return self.bundle or self.binary or self.actions_stamp
 
 
@@ -214,7 +210,7 @@ class Target(object):
 #   to the input file name as well as the output target name.
 
 
-class NinjaWriter(object):
+class NinjaWriter:
     def __init__(
         self,
         hash_for_rules,
@@ -228,11 +224,11 @@ class NinjaWriter(object):
         toplevel_dir=None,
     ):
         """
-    base_dir: path from source root to directory containing this gyp file,
-              by gyp semantics, all input paths are relative to this
-    build_dir: path from source root to build output
-    toplevel_dir: path to the toplevel directory
-    """
+        base_dir: path from source root to directory containing this gyp file,
+                  by gyp semantics, all input paths are relative to this
+        build_dir: path from source root to build output
+        toplevel_dir: path to the toplevel directory
+        """
 
         self.hash_for_rules = hash_for_rules
         self.target_outputs = target_outputs
@@ -263,10 +259,10 @@ class NinjaWriter(object):
     def ExpandSpecial(self, path, product_dir=None):
         """Expand specials like $!PRODUCT_DIR in |path|.
 
-    If |product_dir| is None, assumes the cwd is already the product
-    dir.  Otherwise, |product_dir| is the relative path to the product
-    dir.
-    """
+        If |product_dir| is None, assumes the cwd is already the product
+        dir.  Otherwise, |product_dir| is the relative path to the product
+        dir.
+        """
 
         PRODUCT_DIR = "$!PRODUCT_DIR"
         if PRODUCT_DIR in path:
@@ -303,9 +299,9 @@ class NinjaWriter(object):
 
     def GypPathToNinja(self, path, env=None):
         """Translate a gyp path to a ninja path, optionally expanding environment
-    variable references in |path| with |env|.
+        variable references in |path| with |env|.
 
-    See the above discourse on path conversions."""
+        See the above discourse on path conversions."""
         if env:
             if self.flavor == "mac":
                 path = gyp.xcode_emulation.ExpandEnvVars(path, env)
@@ -324,11 +320,11 @@ class NinjaWriter(object):
     def GypPathToUniqueOutput(self, path, qualified=True):
         """Translate a gyp path to a ninja path for writing output.
 
-    If qualified is True, qualify the resulting filename with the name
-    of the target.  This is necessary when e.g. compiling the same
-    path twice for two separate output targets.
+        If qualified is True, qualify the resulting filename with the name
+        of the target.  This is necessary when e.g. compiling the same
+        path twice for two separate output targets.
 
-    See the above discourse on path conversions."""
+        See the above discourse on path conversions."""
 
         path = self.ExpandSpecial(path)
         assert not path.startswith("$"), path
@@ -361,9 +357,9 @@ class NinjaWriter(object):
 
     def WriteCollapsedDependencies(self, name, targets, order_only=None):
         """Given a list of targets, return a path for a single file
-    representing the result of building all the targets or None.
+        representing the result of building all the targets or None.
 
-    Uses a stamp file if necessary."""
+        Uses a stamp file if necessary."""
 
         assert targets == [item for item in targets if item], targets
         if len(targets) == 0:
@@ -377,14 +373,14 @@ class NinjaWriter(object):
 
     def _SubninjaNameForArch(self, arch):
         output_file_base = os.path.splitext(self.output_file_name)[0]
-        return "%s.%s.ninja" % (output_file_base, arch)
+        return f"{output_file_base}.{arch}.ninja"
 
     def WriteSpec(self, spec, config_name, generator_flags):
         """The main entry point for NinjaWriter: write the build rules for a spec.
 
-    Returns a Target object, which represents the output paths for this spec.
-    Returns None if there are no outputs (e.g. a settings-only 'none' type
-    target)."""
+        Returns a Target object, which represents the output paths for this spec.
+        Returns None if there are no outputs (e.g. a settings-only 'none' type
+        target)."""
 
         self.config_name = config_name
         self.name = spec["target_name"]
@@ -418,20 +414,17 @@ class NinjaWriter(object):
         if self.flavor == "mac":
             self.archs = self.xcode_settings.GetActiveArchs(config_name)
             if len(self.archs) > 1:
-                self.arch_subninjas = dict(
-                    (
-                        arch,
-                        ninja_syntax.Writer(
-                            OpenOutput(
-                                os.path.join(
-                                    self.toplevel_build, self._SubninjaNameForArch(arch)
-                                ),
-                                "w",
-                            )
-                        ),
+                self.arch_subninjas = {
+                    arch: ninja_syntax.Writer(
+                        OpenOutput(
+                            os.path.join(
+                                self.toplevel_build, self._SubninjaNameForArch(arch)
+                            ),
+                            "w",
+                        )
                     )
                     for arch in self.archs
-                )
+                }
 
         # Compute predepends for all rules.
         # actions_depends is the dependencies this target depends on before running
@@ -558,7 +551,7 @@ class NinjaWriter(object):
 
     def _WinIdlRule(self, source, prebuild, outputs):
         """Handle the implicit VS .idl rule for one source file. Fills |outputs|
-    with files that are generated."""
+        with files that are generated."""
         outdir, output, vars, flags = self.msvs_settings.GetIdlBuildData(
             source, self.config_name
         )
@@ -595,7 +588,7 @@ class NinjaWriter(object):
         self, spec, extra_sources, prebuild, mac_bundle_depends
     ):
         """Write out the Actions, Rules, and Copies steps.  Return a path
-    representing the outputs of these steps."""
+        representing the outputs of these steps."""
         outputs = []
         if self.is_mac_bundle:
             mac_bundle_resources = spec.get("mac_bundle_resources", [])[:]
@@ -638,16 +631,16 @@ class NinjaWriter(object):
     def GenerateDescription(self, verb, message, fallback):
         """Generate and return a description of a build step.
 
-    |verb| is the short summary, e.g. ACTION or RULE.
-    |message| is a hand-written description, or None if not available.
-    |fallback| is the gyp-level name of the step, usable as a fallback.
-    """
+        |verb| is the short summary, e.g. ACTION or RULE.
+        |message| is a hand-written description, or None if not available.
+        |fallback| is the gyp-level name of the step, usable as a fallback.
+        """
         if self.toolset != "target":
             verb += "(%s)" % self.toolset
         if message:
-            return "%s %s" % (verb, self.ExpandSpecial(message))
+            return "{} {}".format(verb, self.ExpandSpecial(message))
         else:
-            return "%s %s: %s" % (verb, self.name, fallback)
+            return f"{verb} {self.name}: {fallback}"
 
     def WriteActions(
         self, actions, extra_sources, prebuild, extra_mac_bundle_resources
@@ -657,7 +650,7 @@ class NinjaWriter(object):
         all_outputs = []
         for action in actions:
             # First write out a rule for the action.
-            name = "%s_%s" % (action["action_name"], self.hash_for_rules)
+            name = "{}_{}".format(action["action_name"], self.hash_for_rules)
             description = self.GenerateDescription(
                 "ACTION", action.get("message", None), name
             )
@@ -706,7 +699,7 @@ class NinjaWriter(object):
                 continue
 
             # First write out a rule for the rule action.
-            name = "%s_%s" % (rule["rule_name"], self.hash_for_rules)
+            name = "{}_{}".format(rule["rule_name"], self.hash_for_rules)
 
             args = rule["action"]
             description = self.GenerateDescription(
@@ -731,7 +724,7 @@ class NinjaWriter(object):
             # must vary per source file.
             # Compute the list of variables we'll need to provide.
             special_locals = ("source", "root", "dirname", "ext", "name")
-            needed_variables = set(["source"])
+            needed_variables = {"source"}
             for argument in args:
                 for var in special_locals:
                     if "${%s}" % var in argument:
@@ -875,7 +868,7 @@ class NinjaWriter(object):
         output = self.GypPathToUniqueOutput("headers.hmap")
         self.xcode_settings.header_map_path = output
         all_headers = map(
-            self.GypPathToNinja, filter(lambda x: x.endswith((".h")), all_sources)
+            self.GypPathToNinja, filter(lambda x: x.endswith(".h"), all_sources)
         )
         variables = [
             ("framework", framework),
@@ -925,11 +918,11 @@ class NinjaWriter(object):
     def WriteMacXCassets(self, xcassets, bundle_depends):
         """Writes ninja edges for 'mac_bundle_resources' .xcassets files.
 
-    This add an invocation of 'actool' via the 'mac_tool.py' helper script.
-    It assumes that the assets catalogs define at least one imageset and
-    thus an Assets.car file will be generated in the application resources
-    directory. If this is not the case, then the build will probably be done
-    at each invocation of ninja."""
+        This add an invocation of 'actool' via the 'mac_tool.py' helper script.
+        It assumes that the assets catalogs define at least one imageset and
+        thus an Assets.car file will be generated in the application resources
+        directory. If this is not the case, then the build will probably be done
+        at each invocation of ninja."""
         if not xcassets:
             return
 
@@ -1047,22 +1040,19 @@ class NinjaWriter(object):
                 spec,
             )
         else:
-            return dict(
-                (
-                    arch,
-                    self.WriteSourcesForArch(
-                        self.arch_subninjas[arch],
-                        config_name,
-                        config,
-                        sources,
-                        predepends,
-                        precompiled_header,
-                        spec,
-                        arch=arch,
-                    ),
+            return {
+                arch: self.WriteSourcesForArch(
+                    self.arch_subninjas[arch],
+                    config_name,
+                    config,
+                    sources,
+                    predepends,
+                    precompiled_header,
+                    spec,
+                    arch=arch,
                 )
                 for arch in self.archs
-            )
+            }
 
     def WriteSourcesForArch(
         self,
@@ -1729,8 +1719,8 @@ class NinjaWriter(object):
 
     def GetPostbuildCommand(self, spec, output, output_binary, is_command_start):
         """Returns a shell command that runs all the postbuilds, and removes
-    |output| if any of them fails. If |is_command_start| is False, then the
-    returned string will start with ' && '."""
+        |output| if any of them fails. If |is_command_start| is False, then the
+        returned string will start with ' && '."""
         if not self.xcode_settings or spec["type"] == "none" or not output:
             return ""
         output = QuoteShellArgument(output, self.flavor)
@@ -1776,8 +1766,8 @@ class NinjaWriter(object):
 
     def ComputeExportEnvString(self, env):
         """Given an environment, returns a string looking like
-        'export FOO=foo; export BAR="${FOO} bar;'
-    that exports |env| to the shell."""
+            'export FOO=foo; export BAR="${FOO} bar;'
+        that exports |env| to the shell."""
         export_str = []
         for k, v in env:
             export_str.append(
@@ -1842,7 +1832,7 @@ class NinjaWriter(object):
             "shared_library",
             "executable",
         ):
-            return "%s%s%s" % (prefix, target, extension)
+            return f"{prefix}{target}{extension}"
         elif type == "none":
             return "%s.stamp" % target
         else:
@@ -1909,8 +1899,8 @@ class NinjaWriter(object):
     ):
         """Write out a new ninja "rule" statement for a given command.
 
-    Returns the name of the new rule, and a copy of |args| with variables
-    expanded."""
+        Returns the name of the new rule, and a copy of |args| with variables
+        expanded."""
 
         if self.flavor == "win":
             args = [
@@ -2147,7 +2137,7 @@ def GetDefaultConcurrentLinks():
 
 def _GetWinLinkRuleNameSuffix(embed_manifest):
     """Returns the suffix used to select an appropriate linking rule depending on
-  whether the manifest embedding is enabled."""
+    whether the manifest embedding is enabled."""
     return "_embed" if embed_manifest else ""
 
 
@@ -2538,10 +2528,12 @@ def GenerateOutputForConfig(target_list, target_dicts, data, params, config_name
             "solink",
             description="SOLINK $lib",
             restat=True,
-            command=mtime_preserving_solink_base % {"suffix": "@$link_file_list"},  # noqa: E501
+            command=mtime_preserving_solink_base
+            % {"suffix": "@$link_file_list"},  # noqa: E501
             rspfile="$link_file_list",
-            rspfile_content=("-Wl,--whole-archive $in $solibs -Wl,"
-                             "--no-whole-archive $libs"),
+            rspfile_content=(
+                "-Wl,--whole-archive $in $solibs -Wl," "--no-whole-archive $libs"
+            ),
             pool="link_pool",
         )
         master_ninja.rule(
@@ -2798,8 +2790,8 @@ def GenerateOutputForConfig(target_list, target_dicts, data, params, config_name
 
         this_make_global_settings = data[build_file].get("make_global_settings", [])
         assert make_global_settings == this_make_global_settings, (
-            "make_global_settings needs to be the same for all targets. %s vs. %s"
-            % (this_make_global_settings, make_global_settings)
+            "make_global_settings needs to be the same for all targets. "
+            f"{this_make_global_settings} vs. {make_global_settings}"
         )
 
         spec = target_dicts[qualified_target]
@@ -2891,7 +2883,7 @@ def PerformBuild(data, configurations, params):
     for config in configurations:
         builddir = os.path.join(options.toplevel_dir, "out", config)
         arguments = ["ninja", "-C", builddir]
-        print("Building [%s]: %s" % (config, arguments))
+        print(f"Building [{config}]: {arguments}")
         subprocess.check_call(arguments)
 
 
