@@ -9,10 +9,7 @@
 import unittest
 import gyp.MSVSSettings as MSVSSettings
 
-try:
-    from StringIO import StringIO  # Python 2
-except ImportError:
-    from io import StringIO  # Python 3
+from io import StringIO
 
 
 class TestSequenceFunctions(unittest.TestCase):
